+++ libs/safekeeper_api/src/models.rs
